web/mindsweeper/source/requirements.txt
flask
flask-session
psycopg2-binary